- Исправления 1.17.13 - Подключили swiftLint, начали работу по причесываванию.
