+++ Tracker/Models/CoreData/TrackerRecordCoreData.swift
@@ -11,13 +11,12 @@ import CoreData
 
 @objc(TrackerRecordCoreData)
 public class TrackerRecordCoreData: NSManagedObject {
-    
+
     @NSManaged public var date: String?
     @NSManaged public var id: UUID?
     @NSManaged public var tracker: TrackerCoreData?
-    
+
     @nonobjc public class func fetchRequest() -> NSFetchRequest<TrackerRecordCoreData> {
         return NSFetchRequest<TrackerRecordCoreData>(entityName: "TrackerRecordCoreData")
     }
-    
 }
