+++ Tracker/Supporting/Colors/Colors.swift
@@ -7,8 +7,7 @@
 
 import UIKit
 
-struct AppColors {
-    
+enum AppColors {
     static let textFieldBackground = UIColor(named: "textFieldBackgroundColor")
     static let background = UIColor(named: "projectBackground")
     static let buttonGray = UIColor(named: "createButtonGrayColor")
@@ -17,21 +16,21 @@ struct AppColors {
     static let tabBarBorder = UIColor(named: "tabBarBorderColor")
     static let switchOn = UIColor(named: "switchOnColor")
     static let switchOff = UIColor(named: "switchOffColor")
-    
     static let buttonBlack = UIColor(named: "buttonBackground")
-    static let buttonTextColor = UIColor(named: "buttonTextColor")
+    static let dateButtonText = UIColor(named: "dateButtonTextColor")
+    static let buttonText = UIColor(named: "buttonTextColor")
     static let dateLabelBackground = UIColor(named: "dateLabelBackground")
     static let createButtonText = UIColor(named: "createButtonTextGray")
     static let selectionCell = UIColor(named: "selectionCell")
-    static let textFieldTextColor = UIColor(named: "textFieldTextColor")
-    static let filterButtonTextColor = UIColor(named: "filterButtonTextColor")
-    static let plusButtonColor = UIColor(named: "plusButtonColor")
+    static let textFieldText = UIColor(named: "textFieldTextColor")
+    static let filterButtonText = UIColor(named: "filterButtonTextColor")
+    static let plusButton = UIColor(named: "plusButtonColor")
     static let trackerCellDoneLabel = UIColor(named: "plusButtonColor")
     static let searchTextFieldPlaceholder = UIColor(named: "searchTextFieldPlaceholder")
 
     static let trackerCellTitleLabel = UIColor(named: "titleLabelColor")
     static let trackerCellEmojiRound = UIColor.white.withAlphaComponent(0.1)
-    static let separatorColor = UIColor(named: "tableSeparatorColor")
+    static let separator = UIColor(named: "tableSeparatorColor")
 
     static let gradient1 =  UIColor(named: "gradient1")?.cgColor
     static let gradient2 =  UIColor(named: "gradient2")?.cgColor
