+++ Tracker/ViewControllers/CreatingNewOneOffTrackerVC/ViewModel/CreatingOneOffTrackerViewModel.swift
@@ -8,53 +8,52 @@
 import Foundation
 
 final class CreatingOneOffTrackerViewModel: CreatingOneOffTrackerViewModelProtocol {
-    
+
     var tableViewRows = [SGen.category]
-    
+
     var arrayOfEmoji = MainHelper.arrayOfEmoji
-    
+
     var arrayOfColors = MainHelper.arrayOfColors
-    
+
     let coreDataManager = TrackerCoreManager.shared
-    
+
     var trackerName: String? {
         didSet {
             isDoneButtonEnable?()
         }
     }
-    
+
     var selectedEmoji: String? {
         didSet {
             isDoneButtonEnable?()
         }
     }
-    
+
     var selectedColor: String? {
         didSet {
             isDoneButtonEnable?()
         }
     }
-    
-    var selectedCategory: String?
-    {
+
+    var selectedCategory: String? {
         didSet {
             isDoneButtonEnable?()
         }
     }
-    
+
     var selectedSchedule: String? {
         didSet {
             isDoneButtonEnable?()
         }
     }
-    
+
     var isDoneButtonEnable: ( () -> Void )?
-        
+
     func getBackToMainScreen() {
         let cancelCreatingTrackerNotification = Notification.Name("cancelCreatingTracker")
         NotificationCenter.default.post(name: cancelCreatingTrackerNotification, object: nil)
     }
-    
+
     func createNewTracker() {
         guard let name = trackerName,
               let category = selectedCategory,
@@ -62,7 +61,7 @@ final class CreatingOneOffTrackerViewModel: CreatingOneOffTrackerViewModelProtoc
               let emoji = selectedEmoji else {
             print("Smth's going wrong here"); return
         }
-        
+
         let newTask = TrackerCategory(header: category,
                                       trackers: [Tracker(id: UUID(),
                                                          name: name,
@@ -72,14 +71,14 @@ final class CreatingOneOffTrackerViewModel: CreatingOneOffTrackerViewModelProtoc
         coreDataManager.createNewTracker(newTracker: newTask)
         getBackToMainScreen()
     }
-    
+
     func isAllFieldsFilled() -> Bool {
         let allFieldsFilled =
-            trackerName != nil &&
-            trackerName != "" &&
-            selectedCategory != nil &&
-            selectedEmoji != nil &&
-            selectedColor != nil
+        trackerName != nil &&
+        trackerName != "" &&
+        selectedCategory != nil &&
+        selectedEmoji != nil &&
+        selectedColor != nil
         return allFieldsFilled
     }
 }
