+++ Tracker/ViewControllers/CreatingNewOneOffTrackerVC/Views/CreatingOneOffTrackerVC+CollectionView.swift
@@ -8,26 +8,32 @@
 import UIKit
 
 // MARK: - UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout
-extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
-    
+extension CreatingOneOffTrackerVC: UICollectionViewDataSource,
+                                   UICollectionViewDelegate,
+                                   UICollectionViewDelegateFlowLayout {
+
     func setupEmojiCollectionView() {
         emojiCollection.dataSource = self
         emojiCollection.delegate = self
         emojiCollection.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "emojiCell")
-        emojiCollection.register(SupplementaryView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
+        emojiCollection.register(SupplementaryView.self,
+                                 forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+                                 withReuseIdentifier: "header")
         emojiCollection.backgroundColor = AppColors.background
         emojiCollection.isScrollEnabled = false
     }
-    
+
     func setupColorsCollectionView() {
         colorsCollection.dataSource = self
         colorsCollection.delegate = self
         colorsCollection.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "colorsCell")
-        colorsCollection.register(SupplementaryView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: "header")
+        colorsCollection.register(SupplementaryView.self,
+                                  forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+                                  withReuseIdentifier: "header")
         colorsCollection.backgroundColor = AppColors.background
         colorsCollection.isScrollEnabled = false
     }
-    
+
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         if collectionView == emojiCollection {
             viewModel.arrayOfEmoji.count
@@ -35,9 +41,10 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
             viewModel.arrayOfColors.count
         }
     }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
+
+    func collectionView(_ collectionView: UICollectionView,
+                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+
         if collectionView == emojiCollection {
             let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "emojiCell", for: indexPath)
             let view = UILabel(frame: cell.contentView.bounds)
@@ -58,7 +65,7 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
             return cell
         }
     }
-    
+
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         if collectionView == emojiCollection {
             let cell = collectionView.cellForItem(at: indexPath)
@@ -75,7 +82,7 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
             viewModel.selectedColor = viewModel.arrayOfColors[indexPath.row]
         }
     }
-    
+
     func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
         if collectionView == emojiCollection {
             let cell = collectionView.cellForItem(at: indexPath)
@@ -85,29 +92,38 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
             cell?.layer.borderWidth = 0
         }
     }
-    
+
     private func colorFromHexToRGB(hexColors: [String]) -> [UIColor] {
         return hexColors.map { UIColor(hex: $0) }
     }
-    
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
         return CGSize(width: 52, height: 52)
     }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
         5
     }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        minimumLineSpacingForSectionAt section: Int) -> CGFloat {
         1
     }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        referenceSizeForHeaderInSection section: Int) -> CGSize {
         CGSize(width: collectionView.frame.width, height: 32)
     }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        viewForSupplementaryElementOfKind kind: String,
+                        at indexPath: IndexPath) -> UICollectionReusableView {
         var id = ""
         switch kind {
         case UICollectionView.elementKindSectionHeader:
@@ -117,11 +133,12 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
         default:
             id = ""
         }
-        
-        guard let view = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: id, for: indexPath) as? SupplementaryView else {
+
+        guard let view = collectionView.dequeueReusableSupplementaryView(
+            ofKind: kind, withReuseIdentifier: id, for: indexPath) as? SupplementaryView else {
             print("We have some problems with header"); return UICollectionReusableView()
         }
-        
+
         if collectionView == emojiCollection {
             view.label.text = "Emoji"
         } else {
@@ -129,9 +146,10 @@ extension CreatingOneOffTrackerVC: UICollectionViewDataSource, UICollectionViewD
         }
         return view
     }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        insetForSectionAt section: Int) -> UIEdgeInsets {
         UIEdgeInsets(top: 24, left: 0, bottom: 0, right: 0)
     }
 }
-
