+++ Tracker/ViewControllers/CreatingNewOneOffTrackerVC/Views/CreatingOneOffTrackerVC+ScrollView.swift
@@ -9,41 +9,41 @@ import UIKit
 
 // MARK: - setupScrollView
 extension CreatingOneOffTrackerVC {
-    
+
     func setupScrollView() {
-        
+
         let screenScrollView = UIScrollView()
-        
+
         view.addSubViews([screenScrollView])
-        
+
         NSLayoutConstraint.activate([
             screenScrollView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             screenScrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
             screenScrollView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
             screenScrollView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor)
         ])
-        
+
         let contentView = UIView()
-        
+
         screenScrollView.addSubViews([contentView])
-        
+
         let hConst = contentView.heightAnchor.constraint(equalTo: screenScrollView.heightAnchor)
         hConst.isActive = true
         hConst.priority = UILayoutPriority(50)
-        
+
         NSLayoutConstraint.activate([
             contentView.topAnchor.constraint(equalTo: screenScrollView.topAnchor),
             contentView.leadingAnchor.constraint(equalTo: screenScrollView.leadingAnchor),
             contentView.trailingAnchor.constraint(equalTo: screenScrollView.trailingAnchor),
             contentView.bottomAnchor.constraint(equalTo: screenScrollView.bottomAnchor),
-            
-            contentView.widthAnchor.constraint(equalTo: screenScrollView.widthAnchor),
+
+            contentView.widthAnchor.constraint(equalTo: screenScrollView.widthAnchor)
         ])
-        
+
         setupContentStack()
-        
+
         contentView.addSubview(contentStackView)
-        
+
         NSLayoutConstraint.activate([
             contentStackView.topAnchor.constraint(equalTo: contentView.topAnchor),
             contentStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
@@ -52,48 +52,54 @@ extension CreatingOneOffTrackerVC {
             contentStackView.widthAnchor.constraint(equalTo: contentView.widthAnchor)
         ])
     }
-    
+
     func setupContentStack() {
-                
+
         let textFieldViewStack = UIStackView()
         textFieldViewStack.axis = .vertical
         textFieldViewStack.spacing = 8
         [trackerNameTextField, exceedLabel].forEach { textFieldViewStack.addArrangedSubview($0) }
-         
+
         let buttonsStack = setupButtonsStack()
-        
-        [textFieldViewStack, contentStackView, tableView, emojiCollection, colorsCollection, buttonsStack].forEach { $0.translatesAutoresizingMaskIntoConstraints = false}
-        
-        [textFieldViewStack, tableView, emojiCollection, colorsCollection, buttonsStack].forEach { contentStackView.addArrangedSubview($0) }
-        
+
+        [textFieldViewStack, contentStackView, tableView,
+         emojiCollection, colorsCollection, buttonsStack].forEach {
+            $0.translatesAutoresizingMaskIntoConstraints = false
+        }
+
+        [textFieldViewStack, tableView, emojiCollection,
+         colorsCollection, buttonsStack].forEach {
+            contentStackView.addArrangedSubview($0)
+        }
+
         NSLayoutConstraint.activate([
             trackerNameTextField.heightAnchor.constraint(equalToConstant: 75),
-            
+
             tableView.topAnchor.constraint(equalTo: textFieldViewStack.bottomAnchor, constant: 24),
             tableView.heightAnchor.constraint(equalToConstant: 75),
-            
+
             emojiCollection.topAnchor.constraint(equalTo: tableView.bottomAnchor),
             emojiCollection.heightAnchor.constraint(equalToConstant: 204),
-            
+
             colorsCollection.topAnchor.constraint(equalTo: emojiCollection.bottomAnchor, constant: 8),
             colorsCollection.heightAnchor.constraint(equalToConstant: 230),
-            
+
             buttonsStack.topAnchor.constraint(equalTo: colorsCollection.bottomAnchor, constant: 16),
             buttonsStack.heightAnchor.constraint(equalToConstant: 60)
 
         ])
     }
-    
+
     func setupButtonsStack() -> UIStackView {
         cancelButton.backgroundColor = .clear
         cancelButton.layer.borderWidth = 1
         cancelButton.layer.borderColor = AppColors.buttonRed?.cgColor
         cancelButton.setTitleColor(AppColors.buttonRed, for: .normal)
         cancelButton.addTarget(self, action: #selector(cancelButtonTapped), for: .touchUpInside)
-        
+
         createButton.setTitleColor(AppColors.createButtonText, for: .normal)
         createButton.addTarget(self, action: #selector(createButtonTapped), for: .touchUpInside)
-        
+
         let stack = UIStackView()
         stack.axis = .horizontal
         stack.distribution = .fillEqually
