+++ Tracker/ViewControllers/CreatingNewOneOffTrackerVC/Views/CreatingOneOffTrackerVC+TableView.swift
@@ -7,20 +7,19 @@
 
 import UIKit
 
-// MARK: -  UITableViewDataSource, UITableViewDelegate
 extension CreatingOneOffTrackerVC: UITableViewDataSource, UITableViewDelegate {
-    
+
     func setupTableView() {
         tableView.dataSource = self
         tableView.delegate = self
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
         tableView.layer.cornerRadius = 10
     }
-    
+
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         viewModel.tableViewRows.count
     }
-    
+
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = UITableViewCell(style: .subtitle, reuseIdentifier: "cell")
         cell.textLabel?.text = viewModel.tableViewRows[indexPath.row]
@@ -29,23 +28,23 @@ extension CreatingOneOffTrackerVC: UITableViewDataSource, UITableViewDelegate {
         cell.detailTextLabel?.textColor = AppColors.buttonGray
         cell.textLabel?.font = .systemFont(ofSize: 17, weight: .regular)
         cell.selectionStyle = .none
-        
+
         let disclosureImage = UIImageView(frame: CGRect(x: 0, y: 0, width: 7, height: 12))
         disclosureImage.image = UIImage(named: "chevron")
         cell.accessoryView = disclosureImage
-        
+
         // Убираем сепаратор у последней ячейки
         if indexPath.row == viewModel.tableViewRows.count - 1 {
             cell.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: .greatestFiniteMagnitude)
         }
-        
+
         return cell
     }
-    
+
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         rowHeight
     }
-    
+
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let viewModel = ChoosingCategoryViewModel()
         let categoryVC = ChoosingCategoryViewController(viewModel: viewModel)
@@ -58,9 +57,8 @@ extension CreatingOneOffTrackerVC: UITableViewDataSource, UITableViewDelegate {
         }
         present(navVC, animated: true)
     }
-    
+
     func tableView(_ tableView: UITableView, didDeselectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
     }
 }
-
