+++ Tracker/ViewControllers/CreatingNewTrackerVC/Views/CreatingNewTrackerVC+TableView.swift
@@ -17,7 +17,7 @@ extension CreatingNewTrackerViewController: UITableViewDataSource, UITableViewDe
         
         tableView.layer.cornerRadius = 10
         tableView.separatorInset = UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 10)
-        tableView.separatorColor = AppColors.separatorColor
+        tableView.separatorColor = AppColors.separator
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
