+++ Tracker/ViewControllers/CreatingNewTrackerVC/Views/CreatingNewTrackerVC+TextField.swift
@@ -11,7 +11,7 @@ import UIKit
 extension CreatingNewTrackerViewController: UITextFieldDelegate {
     
     func textFieldDidChangeSelection(_ textField: UITextField) {
-        textField.textColor = AppColors.textFieldTextColor
+        textField.textColor = AppColors.textFieldText
         viewModel.trackerName = textField.text
     }
