+++ Tracker/ViewControllers/CreatingNewTrackerVC/Views/CreatingNewTrackerViewController.swift
@@ -112,7 +112,7 @@ final class CreatingNewTrackerViewController: UIViewController {
         let button = UIButton()
         button.setTitle(title, for: .normal)
         button.titleLabel?.font = .systemFont(ofSize: 16, weight: .medium)
-        button.setTitleColor(AppColors.buttonTextColor, for: .normal)
+        button.setTitleColor(AppColors.buttonText, for: .normal)
         button.backgroundColor = AppColors.buttonBlack
         button.layer.masksToBounds = true
         button.layer.cornerRadius = 15
@@ -149,7 +149,7 @@ final class CreatingNewTrackerViewController: UIViewController {
         
         trackerNameTextField.rightView = clearTextStack
         trackerNameTextField.textAlignment = .left
-        trackerNameTextField.textColor = AppColors.textFieldTextColor
+        trackerNameTextField.textColor = AppColors.textFieldText
     }
     
     // MARK: - Private Methods
