+++ TrackerCoreData.swift
@@ -11,7 +11,7 @@ import CoreData
 
 @objc(TrackerCoreData)
 public class TrackerCoreData: NSManagedObject {
-    
+
     @NSManaged public var colorHex: String?
     @NSManaged public var emoji: String?
     @NSManaged public var id: UUID?
@@ -19,7 +19,7 @@ public class TrackerCoreData: NSManagedObject {
     @NSManaged public var schedule: String?
     @NSManaged public var category: TrackerCategoryCoreData?
     @NSManaged public var trackerRecord: TrackerRecordCoreData?
-    
+
     @nonobjc public class func fetchRequest() -> NSFetchRequest<TrackerCoreData> {
         return NSFetchRequest<TrackerCoreData>(entityName: "TrackerCoreData")
     }
